Moved the button to configuration tab
